made admin and user dashboard pages
